Allows for non-integer countdowns.
This allows countdowns to build up and add days from time to time.
This is odd behavior, but I like the slice of randomness it adds.

+++ models.py
@@ -9,5 +9,5 @@ class Manga(ndb.Model):
     frequency = ndb.FloatProperty(default=1.0)
     freq_unit = ndb.StringProperty(default='')
     update = ndb.BooleanProperty(default=True)
-    countdown = ndb.IntegerProperty(default=0)
+    countdown = ndb.FloatProperty(default=0.0)
     manga_updates_url = ndb.StringProperty(default='')
